requirements.txt
# Core AI
google-generativeai>=0.8.0

# Data validation & settings
pydantic>=2.0
pydantic-settings>=2.0

# Async database
sqlalchemy>=2.0
aiosqlite>=0.20

# PDF parsing
pdfplumber>=0.10

# Terminal UI
rich>=13.0
readchar>=4.0

# Web cockpit (Phase 2)
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
websockets>=12.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
